// src/api/repositories/entities/UserEntity.ts
export type UserEntity = {
  id: string
  name: string
  email: string
  password?: string
  avatar: string
  Author: []
  Publisher: []
  Book: []
}

export type UserPassEntity = {
  id: string
  name: string
  email: string
  password: string
  avatar: string
  Author: []
  Publisher: []
  Book: []
}